WebSettings: Update Tooling
