+++ TF_Settings_Web/src/TouchFree/Tracking/TrackingManager.ts
