--- TF_Tooling_Web/src/Connection/index.ts
@@ -1,11 +1,0 @@
-import { ConnectionManager } from "./ConnectionManager";
-import { MessageReceiver } from "./MessageReceiver";
-import * as TouchFreeServiceTypes from "./TouchFreeServiceTypes";
-import { ServiceConnection } from "./ServiceConnection";
-
-module.exports = {
-    ConnectionManager: ConnectionManager,
-    MessageReceiver: MessageReceiver,
-    TouchFreeServiceTypes: TouchFreeServiceTypes,
-    ServiceConnection: ServiceConnection
-}
