--- TF_Tooling_Web/src/index.ts
@@ -1,17 +1,0 @@
-import * as Configuration from './Configuration'
-import * as Connection from './Connection'
-import * as Cursors from './Cursors'
-import * as InputControllers from './InputControllers';
-import * as Plugins from './Plugins';
-import * as TouchFreeToolingTypes from './TouchFreeToolingTypes';
-import * as Tracking from './Tracking';
-
-module.exports = {
-    Configuration: Configuration,
-    Connection: Connection,
-    Cursors: Cursors,
-    InputControllers: InputControllers,
-    Plugins: Plugins,
-    TouchFreeToolingTypes: TouchFreeToolingTypes,
-    Tracking: Tracking,
-}
